[upgrade] Remove 'linux-firmware-neptune'
and 'linux-firmware-neptune-rtw-debug' if found. The new debug
package was added in SteamOS 3.2 which will get installed and
remove 'linux-firmware' if a user goes through an operating
system upgrade before a winesapOS upgrade. We now ensure to fix
that. The reorder of Arch Linux repositories also ensures that
this will never be replaced again.

Resolves #395