[setup] Do not leak passwords
Disable debug logging before collecting and using passwords. This
way, setup logs can safely be uploaded publicly for debugging and
troubleshooting purposes.

To be clear: passwords were not logged in winesapOS 3.2.1 or older.
winesapOS 3.3.0 is adding the ability to change various passwords
which is why this is now a concern.