[readme] Add clarification
about the SteamOS and winesapOS comparison.

The 16 GiB flash drive storage we provide is only on the
Performance and Secure images (not Minimal).

Instead of calling the desktop environment 'Plasma', we now include
the 'KDE' company name as it is more recognizable.

Discover now, in additional to Flatpaks, supports Pacman
applications and updates in winesapOS (SteamOS still only supports
Flatpaks).