[build] Do not zip the image
GitHub Actions has a limit on the amount of storage we can use for
artifacts. We are not even using these artifacts. The build is a
hybrid between the performance and minimal image which is not
something we want to support. The latest workflows are failing due
to the zip files taking up all remaining storage space. This
change allows the CI pipeline to report successful builds again.