[install] Install Steam
This is a manual revert of:
d8f1286a829e4f490aaa852bd3e29c95493ffdb1

It is deemed safe to install Steam because it does not include the
bootstrap files. Those only get downloaded when a user launches the
Steam client (which is actually a wrapper script to download the
required files first).

Resolves #648