[repo] Do not build COSMIC DE yet
There are 16 required packages or 21 packages including optional
ones that need to be built from the AUR so this would not have
worked anyways. It also leads to 'command not found' because the
'makepkg_fn' was not used. We'll wait to fully add this once
COSMIC DE in a beta and hosted on the Chaotic AUR.

Related #870