Added more apps follow up (#898)
* [changelog] Add latest changes

Due to merge conflicts, the change log was not recently updated.

* [dual-boot] Rename script

to be prefixed with 'winesapos' instead of 'winesap'.

Also make the file executable.

* [dual-boot] Rename desktop shortcut

to be prefixed with 'winesapos' instead of 'winesap' and to use
all lower case letters for consistency.

Also make it executable as some desktop environments require that
on desktop shortcuts.

* [install][tests] Install dual-boot

desktop shortcut and script.

* Rename os-release file

from 'os-winesapos-release-arch' to 'os-release-winesapos'.

* [contributing] Use consistent version

in the 'os-release-winesapos' that 'VERSION' uses.

* [readme][install][tests] Only install KDE Connect on Plasma

It is a lot more complicated to get it running on other desktop
environments. I briefly tested with Cinnamon and could not get it
to work.

Also move mention of GuestSneezeOSDev to the contributor's section
in the read me file.

* [repo] Do not build 'coolercontrol'

It is already built by the Chaotic AUR repository. We do not need
to build it again in the winesapOS repository.

* [install][tests][setup] Install CoolerControl

for computer fan management.

Resolves #894

* [repo] Do not build COSMIC DE yet

There are 16 required packages or 21 packages including optional
ones that need to be built from the AUR so this would not have
worked anyways. It also leads to 'command not found' because the
'makepkg_fn' was not used. We'll wait to fully add this once
COSMIC DE in a beta and hosted on the Chaotic AUR.

Related #870

* [repo] Do not build Waydroid packages

They are already packaged by the Chaotic AUR repository.

* [setup] Install Waydroid

It is 2.5 GiB so only install it as part of the first-time setup.

* [files] Remove Waydroid services

These were originally designed to provide a Waydroid session. I
could not get that to work and the services are otherwise not
needed as running 'waydroid show-full-ui' will automatically start
the required services.