Build improvements follow up (#920)
* [install] Fix broken elif block

It was missing a 'then' statement.

* [files] Make desktop files executable

The new streaming desktop shortcuts for NVIDIA GeForce now and
Xbox Cloud Gaming will not work unless they are executable.
Otherwise, there is prompt to the user asking if they are use they
want to run it.

* [contributing] List new WINESAPOS_DE options

We now support the i3, KDE Plasma Mobile, and Sway desktop
environments.

* [install] Fix KDE Plasma Mobile install

It would accidently run additional install tasks that are for the
full KDE Plasma desktop. Most, if not all of those, do not apply.

* [tests] Add tests for new DEs

Check to see if the desktop environment or tiling manager is
installed for: i3, KDE Plasma Mobile, and Sway.

* [tests] Check for 'man-db'

This will see if basic tools have been installed which were
previously not tested.

* [repo] Do not install container packages

These are already installed via the 'winesapos-install.sh' script.
This function also does not exist here (it is a variable instead).

* [contributing] Mention that Podman is also supported

for doing OS container builds.

* [tests] Check for 'spice-vdagent'

This is now installed for copy and paste support for GNOME Boxes
and Virtual Machine Manager.

* [install][tests][setup][uninstall] Add new streaming shortcuts

for NVIDIA GeForce Now and Xbox Cloud Gaming.

* [changelog] Mention new changes

* [install] Create winesapOS local dir sooner

so that the streaming shortcuts can be copied over.